changes made, wp almost done, 1 pb to fix

+++ srcs/requirements/wordpress/tools/wordpress_install.sh
@@ -2,20 +2,23 @@
 
 set -e
 
-# Install wordpress
-curl -LO https://wordpress.org/latest.tar.gz
-tar -xvzf latest.tar.gz
-chown -R www-data:www-data wordpress/
-chmod -R 755 wordpress/
-rm latest.tar.gz
+if [ -f ./wordpress/wp-config.php ]
+then
+	echo "Wordpress already configured"
+else
+    # --- Install wordpress ---
+    curl -LO https://wordpress.org/latest.tar.gz
+    tar -xvzf latest.tar.gz
+    chown -R www-data:www-data wordpress/
+    chmod -R 755 wordpress/
+    rm latest.tar.gz
 
-# Set up the wp-config.php file for WordPress
-cp wordpress/wp-config-sample.php wordpress/wp-config.php
-sed -i "s/database_name_here/$MYSQL_DATABASE/" wordpress/wp-config.php
-sed -i "s/username_here/$MYSQL_USER/" wordpress/wp-config.php
-sed -i "s/password_here/$MYSQL_PASSWORD/" wordpress/wp-config.php
-
-# SALTS=$(curl -s https://api.wordpress.org/secret-key/1.1/salt/)
-# sed -i "/# Authentication Unique Keys and Salts/ r /dev/stdin" ./wp-config.php <<< "$SALTS"
+    # --- Setup wp-config.php ---
+    cp wordpress/wp-config-sample.php wordpress/wp-config.php
+    sed -i "s/database_name_here/$MYSQL_DATABASE/" wordpress/wp-config.php
+    sed -i "s/localhost/$MYSQL_HOSTNAME/g" wordpress/wp-config.php
+    sed -i "s/username_here/$MYSQL_USER/" wordpress/wp-config.php
+    sed -i "s/password_here/$MYSQL_PASSWORD/" wordpress/wp-config.php
+fi
 
 exec php-fpm7.4 -F
